Fixing buy/sell issues following a web timeout
This is a very low-tech fix for the issue where the NN suffers some form of web timeout and gets stuck in a loop trying to sell (where gekko is not exposed) or trying to buy (where gekko is already exposed).
All I have done is removed the onTrade function and moved the code to where NN buys or sells. It sells twice so I should probably have created a function for this but as it's only caused twice it doesn't seem that inefficient.
Anyway - it works!
Since I kicked off a bot yesterday it has suffered a web timeout but (unlike before my change) the NN has recovered to continue buying and selling.